+++ src/1_api.js
@@ -8,6 +8,10 @@ goog.require('utils');
 goog.require('goog.json');
 goog.require('storage'); // jshint unused:false
 
+var RETRIES = 2,
+	RETRY_DELAY = 200,
+	TIMEOUT = 5000;
+
 /**
  * @class Server
  * @constructor
@@ -109,7 +113,7 @@ var jsonp_callback_index = 0;
  * @param {string} requestURL
  * @param {Object} requestData
  * @param {utils._httpMethod} requestMethod
- * @param {function(?Error,*=)=} callback
+ * @param {function(?Error,*=,?=)=} callback
  */
 Server.prototype.jsonpRequest = function(requestURL, requestData, requestMethod, callback) {
 	var callbackString = 'branch_callback__' + (this._jsonp_callback_index++);
@@ -119,8 +123,8 @@ Server.prototype.jsonpRequest = function(requestURL, requestData, requestMethod,
 
 	var timeout_trigger = window.setTimeout(function() {
 		window[callbackString] = function() { };
-		callback(new Error(utils.messages.timeout));
-	}, 5000);
+		callback(new Error(utils.messages.timeout), null, 504);
+	}, TIMEOUT);
 
 	window[callbackString] = function(data) {
 		window.clearTimeout(timeout_trigger);
@@ -135,35 +139,35 @@ Server.prototype.jsonpRequest = function(requestURL, requestData, requestMethod,
  * @param {Object} data
  * @param {utils._httpMethod} method
  * @param {BranchStorage} storage
- * @param {function(?Error,*=)=} callback
+ * @param {function(?Error,*=,?=)=} callback
  */
 Server.prototype.XHRRequest = function(url, data, method, storage, callback) {
 	var req = window.XMLHttpRequest ? new XMLHttpRequest() : new ActiveXObject("Microsoft.XMLHTTP");
 	req.ontimeout = function() {
-		callback(new Error(utils.messages.timeout));
+		callback(new Error(utils.messages.timeout), null, 504);
 	};
 	req.onreadystatechange = function() {
 		if (req.readyState === 4) {
 			if (req.status === 200) {
 				try {
-					callback(null, goog.json.parse(req.responseText));
+					callback(null, goog.json.parse(req.responseText), req.status);
 				}
 				catch (e) {
-					callback(null, { });
+					callback(null, { }, req.status);
 				}
 			}
 			else if (req.status === 402) {
-				callback(new Error('Not enough credits to redeem.'));
+				callback(new Error('Not enough credits to redeem.'), null, req.status);
 			}
 			else if (req.status.toString().substring(0, 1) === "4" || req.status.toString().substring(0, 1) === "5") {
-				callback(new Error('Error in API: ' + req.status));
+				callback(new Error('Error in API: ' + req.status), null, req.status);
 			}
 		}
 	};
 
 	try {
 		req.open(method, url, true);
-		req.timeout = 5000;
+		req.timeout = TIMEOUT;
 		req.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded');
 		req.send(data);
 	}
@@ -180,6 +184,9 @@ Server.prototype.XHRRequest = function(url, data, method, storage, callback) {
  * @param {function(?Error,*=)=} callback
  */
 Server.prototype.request = function(resource, data, storage, callback) {
+	var self = this;
+
+
 	var u = this.getUrl(resource, data);
 	if (u.error) { return callback(new Error(u.error)); }
 
@@ -191,10 +198,31 @@ Server.prototype.request = function(resource, data, storage, callback) {
 		url = u.url;
 		postData = u.data;
 	}
-	if (storage['getItem']('use_jsonp') || resource.jsonp) {
-		this.jsonpRequest(url, data, resource.method, callback);
-	}
-	else {
-		this.XHRRequest(url, postData, resource.method, storage, callback);
-	}
+
+	// How many times to retry the request if the initial attempt fails
+	var retries = RETRIES;
+	// If request fails, retry after X miliseconds
+	/***
+	 * @type {function(?Error,*=): ?undefined}
+	 */
+	var done = function(err, data, status) {
+		if (err && retries > 0 && status.toString().substring(0, 1) === "5") {
+			retries--;
+			window.setTimeout(function() {
+				makeRequest();
+			}, RETRY_DELAY);
+		}
+		else {
+			callback(err, data);
+		}
+	};
+	var makeRequest = function() {
+		if (storage['getItem']('use_jsonp') || resource.jsonp) {
+			self.jsonpRequest(url, data, resource.method, done);
+		}
+		else {
+			self.XHRRequest(url, postData, resource.method, storage, done);
+		}
+	};
+	makeRequest();
 };