+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The Branch Web SDK adheres to [Semantic Versioning](http://semver.org/).
 
 ## [VERSION] - unreleased
 - Fixes bug with iOS and Android app banner not opening
+### Improvements
+- Behind the scenes improvement to retry all XHR and JSONP requests 3 times if they fail or timeout
 
 ## [v1.5.3] - 2015-05-21
 ### New Functionality